Correct number of skills and tools, add suggested skills
I noticed that some of the Flexible Backgrounds did not correctly match the number of skills and/or tools actually granted by the background in the base game. I also added blank features to display the original defaults as "Suggested Skills". This preserves the intended flavor of the background as written and also is helpful for those of us who mainly just use it to duplicate the D&D rule that you can choose a different skill or tool if your background grants one you already have from another source.