Update PrayerOfHealing.json
Healing has been buffed in one D&D.
I think the buffed content doesn't seem to fit Solastar, so I think it needs to be buffed in a different way.
The short rest effect is not suitable even in the Solasta environment, where short rest can be used at any time. So it seems right to increase the healing amount.

Signed-off-by: kdgmo <135420589+kdgmo@users.noreply.github.com>