Update MassHealingWord.json
Healing has been buffed in one D&D.
However, considering that it is a level 3 spell, oneD&D's buff has little effect, so it feels like a little more buffing is needed. So wouldn't it be correct to increase the recovery dice by 1 and increase the upcasting effect dice by 1? Considering that the dice itself is a small spell, I think a healing amount of about 1/4 of the fireball value is correct.

Signed-off-by: kdgmo <135420589+kdgmo@users.noreply.github.com>